validation and all methods complete

+++ HealthAPI/HealthAPI.Provider/Providers/PatientProvider.cs
@@ -17,11 +17,13 @@ namespace HealthAPI.Provider.Providers
     {
         private readonly ILogger<PatientProvider> _logger;
         private readonly IPatientRepository _patientRepository;
+        private readonly IEncounterRepository _encounterRepository;
 
-        public PatientProvider(IPatientRepository patientRepository, ILogger<PatientProvider> logger)
+        public PatientProvider(IPatientRepository patientRepository, IEncounterRepository encounterRepository, ILogger<PatientProvider> logger)
         {
             _logger = logger;
             _patientRepository = patientRepository;
+            _encounterRepository = encounterRepository;
         }
 
         /// <summary>
@@ -34,13 +36,22 @@ namespace HealthAPI.Provider.Providers
             ValidatePatientInputFields(newPatient);
 
             Patient savedPatient;
+            Patient uniquePatient;
 
-            // Movie email validation to repository
-            //if (await ValidateSingleSku(newMovie))
-            //{
-            //    _logger.LogError("SKU is already in use. Please enter new SKU.");
-            //    throw new ConflictException("SKU is already in use. Please enter new SKU.");
-            //}
+            try
+            {
+                uniquePatient = await _patientRepository.GetPatientByEmailAsync(newPatient);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                throw new ServiceUnavailableException("There was a problem connecting to the database.");
+            }
+
+            if (uniquePatient != null || uniquePatient != default)
+            {
+                throw new ConflictException("Email address already exists in the system. Please enter a unique email address.");
+            }
 
             try
             {
@@ -135,26 +146,41 @@ namespace HealthAPI.Provider.Providers
                 throw new NotFoundException($"Patient with id of {id} not found.");
             }
 
-            // MOVE EMAIL VALIDATION TO REPOSITORY
-            //if (updatedMovie.Sku != existingMovie.Sku)
-            //{
-            //    if (await ValidateSingleSku(updatedMovie))
+            Patient uniquePatientEmail;
 
-            //    {
-            //        _logger.LogError("SKU is already in use. Please enter new SKU.");
-            //        throw new ConflictException("SKU is already in use. Please enter new SKU.");
-            //    }
-            //}
+            if (updatedPatient.Email != existingPatient.Email)
+            {
+                try
+                {
+                    uniquePatientEmail = await _patientRepository.GetPatientByEmailAsync(updatedPatient);
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex.Message);
+                    throw new ServiceUnavailableException("There was a problem connecting to the database.");
+                }
 
-            if (updatedPatient.Id == default)
-                updatedPatient.Id = id;
+                if (uniquePatientEmail != null || uniquePatientEmail != default)
+                {
+                    throw new ConflictException("Email address already exists in the system. Please enter a unique email address.");
+                }
+            }
+
+            if (existingPatient.Id != updatedPatient.Id)
+            {
+                if (updatedPatient.Id == default)
+                {
+                    updatedPatient.Id = existingPatient.Id;
+                }
+                else throw new BadRequestException("Patient ID cannot be changed.");
+            }
 
             ValidatePatientInputFields(updatedPatient);
 
             try
             {
                 await _patientRepository.UpdatePatientAsync(updatedPatient);
-                _logger.LogInformation("Movie updated.");
+                _logger.LogInformation("Patient updated.");
             }
             catch (Exception ex)
             {
@@ -173,23 +199,25 @@ namespace HealthAPI.Provider.Providers
 
         public async Task<Patient> DeletePatientByIdAsync(int patientId)
         {
-
-            var patient = await _patientRepository.DeletePatientByIdAsync(patientId);
-            
-            if (patient == null || patient == default)
+            if (await IsDeleteable(patientId))
             {
-                _logger.LogInformation($"Patient with id: {patientId} could not be found.");
-                throw new NotFoundException($"Patient with id: {patientId} could not be found.");
+                var patient = await _patientRepository.DeletePatientByIdAsync(patientId);
+                if (patient == null || patient == default)
+                {
+                    _logger.LogInformation($"Patient with id: {patientId} could not be found.");
+                    throw new NotFoundException($"Patient with id: {patientId} could not be found.");
+                }
+                return patient;
             }
-            return patient;
+            throw new ConflictException("Patient has associated encounters and cannot be deleted.");
         }
 
         ///Validation Methods
 
         /// <summary>
-        /// Validates a movie object's input fields based on various criteria. Returns list of exceptions if any exist.
+        /// Validates a patient object's input fields based on various criteria. Returns list of exceptions if any exist.
         /// </summary>
-        /// <param name = "newMovie" ></ param >
+        /// <param name = "newPatient" ></ param >
 
         public void ValidatePatientInputFields(Patient newPatient)
         {
@@ -199,17 +227,33 @@ namespace HealthAPI.Provider.Providers
             {
                 patientExceptions.Add("Patient first name is required.");
             }
+            else if (!ValidateAlphabeticNameInput(newPatient.FirstName))
+            {
+                patientExceptions.Add("Patient first name must contain only alphabetic characters.");
+            }
             if (ValidateIfEmptyOrNull(newPatient.LastName))
             {
                 patientExceptions.Add("Patient last name is required.");
+            }
+            else if (!ValidateAlphabeticNameInput(newPatient.LastName))
+            {
+                patientExceptions.Add("Patient first name must contain only alphabetic characters.");
             }
             if (ValidateIfEmptyOrNull(newPatient.Ssn))
             {
                 patientExceptions.Add("Social security number is required.");
             }
+            else if (!ValidateSsnFormat(newPatient.Ssn))
+            {
+                patientExceptions.Add("SSN must match format xxx-xx-xxxx.");
+            }
             if (ValidateIfEmptyOrNull(newPatient.Email))
             {
                 patientExceptions.Add("Email is required.");
+            }
+            else if (!ValidateEmailFormat(newPatient.Email))
+            {
+                patientExceptions.Add("Email must match format x@x.x.");
             }
             if (ValidateIfEmptyOrNull(newPatient.Street))
             {
@@ -223,21 +267,41 @@ namespace HealthAPI.Provider.Providers
             {
                 patientExceptions.Add("State is required.");
             }
+            else if (!ValidateStateFormat(newPatient.State))
+            {
+                patientExceptions.Add("State must be valid two-character state code (eg. AL)");
+            }
             if (ValidateIfEmptyOrNull(newPatient.Postal))
             {
                 patientExceptions.Add("Postal code is required.");
+            }
+            else if (!ValidateZipFormat(newPatient.Postal))
+            {
+                patientExceptions.Add("Postal code must be 5 or 9 digits.");
             }
             if (ValidateIfEmptyOrNull(newPatient.Age.ToString()))
             {
                 patientExceptions.Add("Patient age is required.");
             }
+            else if (!ValidateWholeInteger(newPatient.Age.ToString()))
+            {
+                patientExceptions.Add("Age input must only contain numeric digits.");
+            }
             if (ValidateIfEmptyOrNull(newPatient.Height.ToString()))
             {
                 patientExceptions.Add("Patient height is required.");
             }
+            else if (!ValidateWholeInteger(newPatient.Height.ToString()))
+            {
+                patientExceptions.Add("Height input must only contain numeric digits.");
+            }
             if (ValidateIfEmptyOrNull(newPatient.Weight.ToString()))
             {
-                patientExceptions.Add("Pateint weight is required.");
+                patientExceptions.Add("Patient weight is required.");
+            }
+            else if (!ValidateWholeInteger(newPatient.Weight.ToString()))
+            {
+                patientExceptions.Add("Weight input must only contain numeric digits.");
             }
             if (ValidateIfEmptyOrNull(newPatient.Insurance))
             {
@@ -247,14 +311,10 @@ namespace HealthAPI.Provider.Providers
             {
                 patientExceptions.Add("Patient gender is required.");
             }
-            //else
-            //{
-            //    if (!ValidatePriceFormat(newMovie.DailyRentalCost.ToString()))
-            //        movieExceptions.Add("Price value must match the format DD.DD, where D is any number.");
-
-            //}
-
-
+            else if (!ValidateGenderInput(newPatient.Gender))
+            {
+                patientExceptions.Add("Patient gender must be 'Male', 'Female', or 'Other'.");
+            }
 
             if (patientExceptions.Count > 0)
             {
@@ -273,50 +333,100 @@ namespace HealthAPI.Provider.Providers
             return string.IsNullOrWhiteSpace(modelField);
         }
 
-        ///// <summary>
-        ///// Validatation method to check if the price format of a movie object is correctly formated.
-        ///// </summary>
-        ///// <param name="modelField"></param>
-        ///// <returns>Boolean</returns>
-        //public bool ValidatePriceFormat(string modelField)
-        //{
-        //    Regex priceFormat = new Regex(@"\d+\.\d\d(?!\d)");
-        //    return priceFormat.IsMatch(modelField);
-        //}
+        /// <summary>
+        /// Validates if a string only contains alphabetic name characters.
+        /// Allows spaces, hyphens, and apostrophes. Used to validate patient first and last names.
+        /// </summary>
+        /// <param name="modelField">string name input field</param>
+        /// <returns>boolean, true if input matches regex</returns>
+        public bool ValidateAlphabeticNameInput(string modelField)
+        {
+            var nameCheck = new Regex(@"^[\p{L} \'\-]+$");
+            return nameCheck.IsMatch(modelField);
+        }
 
-        ///// <summary>
-        ///// Validation method to check if the SKU format of a movie object is correctly formatted
-        ///// </summary>
-        ///// <param name="modelField"></param>
-        ///// <returns>Boolean</returns>
-        //public bool ValidateSkuFormat(string modelField)
-        //{
-        //    Regex skuFormat = new Regex(@"^[A-Z]{6}[-][0-9]{4}$");
-        //    return skuFormat.IsMatch(modelField);
 
-        //}
+        /// <summary>
+        /// Validates if a social security input matches format DDD-DD-DDDD
+        /// </summary>
+        /// <param name="modelField">string ssn input field</param>
+        /// <returns>boolean, true if input matches regex</returns>
+        public bool ValidateSsnFormat(string modelField)
+        {
+            var ssnCheck = new Regex(@"^[\d]{3}[-][\d]{2}[-][\d]{4}$");
+            return ssnCheck.IsMatch(modelField);
+        }
 
-        ///// <summary>
-        ///// Validation method to check if the SKU a movie object is currently in use
-        ///// </summary>
-        ///// <param name="addedMovie"></param>
-        ///// <returns>Boolean</returns>
-        //public async Task<bool> ValidateSingleSku(Movie addedMovie)
-        //{
-        //    var movieList = await GetAllMoviesAsync();
-        //    var movieSkuList = new List<string>();
-        //    foreach (var movie in movieList)
-        //    {
-        //        movieSkuList.Add(movie.Sku);
-        //    }
+        /// <summary>
+        /// Validates if user email address is formatted correctly (user@email.com) with alphanumeric username and alphatic domain (EMAIL)
+        /// </summary>
+        /// <param name="emailAddress">string email input field</param>
+        /// <returns>boolean, true if the email is formatted correctly</returns>
+        public bool ValidateEmailFormat(string emailAddress)
+        {
+            var emailCheck = new Regex(@"^\w+@([a-z]+\.)+[a-z]+$");
+            return emailCheck.IsMatch(emailAddress);
+        }
 
-        //    if (movieSkuList.Contains(addedMovie.Sku))
-        //    {
-        //        return true;
-        //    }
-        //    return false;
-        //}
+        /// <summary>
+        /// Validates if user state is formatted correctly (LL). Only accepts true US State and territory abbreviations.
+        /// </summary>
+        /// <param name="state">string email input field</param>
+        /// <returns>boolean, true if the email is formatted correctly</returns>
+        public bool ValidateStateFormat(string state)
+        {
+            var stateCheck = new Regex(@"^((A[LKSZR])|(C[AOT])|(D[EC])|(F[ML])|(G[AU])|(HI)|(I[DLNA])|(K[SY])|(LA)|(M[EHDAINSOT])|(N[EVHJMYCD])|(MP)|(O[HKR])|(P[WAR])|(RI)|(S[CD])|(T[NX])|(UT)|(V[TIA])|(W[AVIY]))$");
+            return stateCheck.IsMatch(state);
+        }
 
+
+        /// <summary>
+        /// Validates if patient postal code is only five or nine numerical digits (ZIPCODE)
+        /// </summary>
+        /// <param name="zipCode">string zipcode input field</param>
+        /// <returns>boolean, true if the zipcode is valid</returns>
+        public bool ValidateZipFormat(string zipCode)
+        {
+            var zipCodeCheck = new Regex(@"^[0-9]{5}(?:-[0-9]{4})?$");
+            return zipCodeCheck.IsMatch(zipCode);
+        }
+
+        /// <summary>
+        /// Validates if input contains only numeric digits, used to validate patient height, weight, and age.
+        /// </summary>
+        /// <param name="input">string input field</param>
+        /// <returns>boolean, true if the input is valid</returns>
+        public bool ValidateWholeInteger(string input)
+        {
+            var integerCheck = new Regex(@"^\d+$");
+            return integerCheck.IsMatch(input);
+        }
+
+        /// <summary>
+        /// Validates if gender input matches either "Male", "Female", or "Other"
+        /// </summary>
+        /// <param name="input">string input field</param>
+        /// <returns>boolean, true if the input is valid</returns>
+        public bool ValidateGenderInput(string input)
+        {
+            var genderCheck = new Regex(@"^(?:male|Male|female|Female|other|Other)$");
+            return genderCheck.IsMatch(input);
+        }
+
+        /// <summary>
+        /// Checks if patient has associated encounters in database.
+        /// </summary>
+        /// <param name="patientId">string input field</param>
+        /// <returns>boolean, true if no encounters exist</returns>
+        public async Task<bool> IsDeleteable(int patientId)
+        {
+            var encounters = await _encounterRepository.GetAllEncountersByIdAsync(patientId);
+            if (encounters.Count == 0)
+            {
+                return true;
+            }
+            return false;
+        }
     }
 
 }
